requirements.txt
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dotenv==0.19.2
